ajustando codigo auditorias

+++ Auditorias Conector/Models/DAO/AuditoriaDAO.cs
@@ -113,7 +113,7 @@ public class Producto
 {
     public string Concepto { get; set; }
     public int Cantidad { get; set; }
-    public int Monto { get; set; }
+    public float Monto { get; set; }
     public string ProductoCodigo { get; set; }
 }
